+++ trafaret/internet.py
@@ -1,0 +1,132 @@
+# -*- coding: utf-8 -*-
+
+import re
+from .regexp import Regexp
+from .base import DataError, String, Bytes, OnError
+from .lib import py3
+
+if py3:
+    import urllib.parse as urlparse
+else:
+    import urlparse
+
+
+MAX_EMAIL_LEN = 254
+
+EMAIL_REGEXP = re.compile(
+    r"(?P<name>^[-!#$%&'*+/=?^_`{}|~0-9A-Z]+(\.[-!#$%&'*+/=?^_`{}|~0-9A-Z]+)*"  # dot-atom
+    r'|^"([\001-\010\013\014\016-\037!#-\[\]-\177]|\\[\001-011\013\014\016-\177])*"' # quoted-string
+    r')@(?P<domain>(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}\.?|[A-Z0-9-]{2,}\.?)$)'  # domain
+    r'|\[(25[0-5]|2[0-4]\d|[0-1]?\d?\d)(\.(25[0-5]|2[0-4]\d|[0-1]?\d?\d)){3}\]$',  # literal form, ipv4 address (SMTP 4.1.3)
+    re.IGNORECASE,
+)
+
+def email_idna_encode(value):
+    if '@' in value:
+        parts = value.split('@')
+        try:
+            parts[-1] = parts[-1].encode('idna').decode('ascii')
+            return '@'.join(parts)
+        except UnicodeError:
+            pass
+    return value
+
+email_regexp_trafaret = OnError(Regexp(EMAIL_REGEXP), 'value is not a valid email address')
+email_trafaret = email_regexp_trafaret | ((Bytes('utf-8') | String()) & email_idna_encode & email_regexp_trafaret)
+Email = lambda: String(allow_blank=True) & OnError(
+    String(max_length=MAX_EMAIL_LEN) & email_trafaret,
+    'value is not a valid email address',
+)
+
+
+URL_REGEXP = re.compile(
+    r'^(?:http|ftp)s?://'  # http:// or https://
+    r'(?:\S+(?::\S*)?@)?'  # user and password
+    r'(?:(?:[A-Z0-9](?:[A-Z0-9-_]{0,61}[A-Z0-9])?\.)+(?:[A-Z]{2,6}\.?|[A-Z0-9-]{2,}\.?)|'  # domain...
+    r'localhost|'  # localhost...
+    r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})'  # ...or ip
+    r'(?::\d+)?'  # optional port
+    r'(?:/?|[/?]\S+)$',
+    re.IGNORECASE,
+)
+URLRegexp = OnError(Regexp(URL_REGEXP), 'value is not URL')
+URLRegexp = Regexp(URL_REGEXP)
+
+def decode_url_idna(value):
+    try:
+        scheme, netloc, path, query, fragment = urlparse.urlsplit(value)
+        netloc = netloc.encode('idna').decode('ascii') # IDN -> ACE
+    except UnicodeError: # invalid domain part
+        pass
+    else:
+        return urlparse.urlunsplit((scheme, netloc, path, query, fragment))
+    return value
+
+URL = lambda: OnError(
+    URLRegexp | ((Bytes('utf-8') | String()) & decode_url_idna & URLRegexp),
+    'value is not URL',
+)
+
+
+class IPv4(Regexp):
+    """
+    >>> IPv4().check('127.0.0.1')
+    '127.0.0.1'
+    """
+
+    regex = re.compile(
+        r'^((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])$',  # noqa
+    )
+
+    def __init__(self):
+        super(IPv4, self).__init__(self.regex)
+
+    def check_and_return(self, value):
+        try:
+            return super(IPv4, self).check_and_return(value)
+        except DataError:
+            self._failure('value is not IPv4 address')
+
+    def __repr__(self):
+        return '<IPv4>'
+
+
+class IPv6(Regexp):
+    """
+    >>> IPv6().check('2001:0db8:0000:0042:0000:8a2e:0370:7334')
+    '2001:0db8:0000:0042:0000:8a2e:0370:7334'
+    """
+
+    regex = re.compile(
+        r'^('
+        r'(::)|'
+        r'(::[0-9a-f]{1,4})|'
+        r'([0-9a-f]{1,4}:){7,7}[0-9a-f]{1,4}|'
+        r'([0-9a-f]{1,4}:){1,7}:|'
+        r'([0-9a-f]{1,4}:){1,6}:[0-9a-f]{1,4}|'
+        r'([0-9a-f]{1,4}:){1,5}(:[0-9a-f]{1,4}){1,2}|'
+        r'([0-9a-f]{1,4}:){1,4}(:[0-9a-f]{1,4}){1,3}|'
+        r'([0-9a-f]{1,4}:){1,3}(:[0-9a-f]{1,4}){1,4}|'
+        r'([0-9a-f]{1,4}:){1,2}(:[0-9a-f]{1,4}){1,5}|'
+        r'[0-9a-f]{1,4}:((:[0-9a-f]{1,4}){1,6})|'
+        r':((:[0-9a-f]{1,4}){1,7}:)|'
+        r'fe80:(:[0-9a-f]{0,4}){0,4}%[0-9a-z]{1,}|'
+        r'::(ffff(:0{1,4}){0,1}:){0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|'  # noqa
+        r'([0-9a-f]{1,4}:){1,4}:((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])'  # noqa
+        r')$',
+        re.IGNORECASE,
+    )
+
+    def __init__(self):
+        super(IPv6, self).__init__(self.regex)
+
+    def check_and_return(self, value):
+        try:
+            return super(IPv6, self).check_and_return(value)
+        except DataError:
+            self._failure('value is not IPv6 address')
+
+    def __repr__(self):
+        return '<IPv6>'
+
+IP = OnError(IPv4 | IPv6, 'value is not IP address')
